Revert & update on location geoname
