+++ app/models/solr_document.rb
@@ -82,7 +82,7 @@ class SolrDocument
   end
 
   def based_near_linked_label
-    ScholarsArchive::LocationLabelParserService.location_parse_uris(ScholarsArchive::LabelParserService.parse_label_uris(self['based_near_linked_ssim']))
+    ScholarsArchive::LabelParserService.location_parse_uris(self['based_near_linked_ssim'])
   end
 
   def embargo_date_range
