+++ spec/workers/fetch_graph_worker_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe FetchGraphWorker, type: :worker do
         allow_any_instance_of(Hyrax::ControlledVocabularies::Location).to receive(:solrize).and_return(['https://sws.geonames.org/5761960/', { label: 'Yamhill County$https://sws.geonames.org/5761960/' }])
         stub_request(:get, 'https://sws.geonames.org/5761960/').with(headers: { 'Accept' => 'application/n-triples, text/plain;q=0.2, application/ld+json, application/x-ld+json, application/rdf+xml, text/turtle, text/rdf+turtle, application/turtle;q=0.2, application/x-turtle;q=0.2, application/n-quads, text/x-nquads;q=0.2, application/n-quads, text/x-nquads;q=0.2, application/rdf+json, text/html;q=0.5, application/xhtml+xml;q=0.7, image/svg+xml;q=0.4, text/n3, text/rdf+n3;q=0.2, application/rdf+n3;q=0.2, application/normalized+n-quads, application/x-normalized+n-quads, text/csv;q=0.4, text/tab-separated-values;q=0.4, application/csvm+json, application/trig, application/x-trig;q=0.2, application/trix', 'Accept-Encoding' => 'gzip;q=1.0,deflate;q=0.6,identity;q=0.3', 'Host' => 'sws.geonames.org', 'User-Agent' => 'Ruby RDF.rb/3.2.4' }).to_return(status: 200, body: '', headers: {})
         stub_request(:get, 'http://ci-test:8080/bigdata/namespace/rw/sparql?GETSTMTS&includeInferred=false&s=%3Chttps://sws.geonames.org/5761960/%3E').with(headers: { 'Accept' => '*/*', 'Accept-Encoding' => 'gzip;q=1.0,deflate;q=0.6,identity;q=0.3', 'Connection' => 'keep-alive', 'Host' => 'ci-test:8080', 'Keep-Alive' => '30', 'User-Agent' => 'Ruby' }).to_return(status: 200, body: '', headers: {})
-        allow(ScholarsArchive::LocationLabelParserService).to receive(:location_parse_labels).and_return(['Yamhill County'])
+        allow(ScholarsArchive::LabelParserService).to receive(:location_parse_labels).and_return(['Yamhill County'])
       end
 
       it 'fetches a work and indexes its linked data labels' do
